+++ CHANGES.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Version: 2.55
+
+- [feature] Implement Venus#date
+- [feature] Implement Venus#match
+- [feature] Implement Venus#gather
+- [update] Make Venus::Unpack#{checks,validate} list context aware
+- [update] Resolve Venus#unpack signature/caller issue
+- [update] Update arguments passing in Venus#{json,perl,yaml}
+- [update] Update {json,perl,yaml} behaviors and error conditions
+
 ## Version: 2.50
 
 - [feature] Implement Venus#work